+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Todos los cambios notables(?) del proyecto serán documentados en este archivo.
 ## [SIN PUBLICAR]
 
 ### 2019-03-24
+- agregado /MOTD /IPToNick /GRABAR /GUARDAMAPA /HABILITAR
 - agregados /PING /FORCEMIDI /FORCEWAV /FORCEMIDIMAP /FORCEWAVMAP /TELEPLOC /NICK2IP /_BUG
 - agregado Mover Inventario del Banco.